+++ codec/doc.go
@@ -1,0 +1,3 @@
+// Package codec provides an implementation of [marshalkit.Marshaler] that uses
+// a set of priority-ordered codecs to marshal and unmarshal types and values.
+package codec